requirements.txt
click==8.1.3
numpy>=1.21.0,<2.0.0
requests==2.28.1
plotext
matplotlib==3.6.2
python-dotenv==0.21.0
PyGithub==1.58.2